requirements.txt
alembic==1.16.2
annotated-types==0.7.0
antlr4-python3-runtime==4.13.2
anyio==4.9.0
attrs==25.3.0
backoff==2.2.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
click-aliases==1.0.5
coloredlogs==14.0
croniter==6.0.0
cryptography==45.0.4
dagster==1.11.0
dagster-cloud-cli==1.11.0
dagster-dg-cli==1.11.0
dagster-dg-core==1.11.0
dagster-graphql==1.11.0
dagster-pipes==1.11.0
dagster-webserver==1.11.0
dagster_shared==1.11.0
dnspython==2.7.0
docstring_parser==0.16
filelock==3.18.0
fsspec==2025.5.1
github3.py==4.0.1
gql==3.5.3
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
grpcio==1.73.1
grpcio-health-checking==1.62.3
h11==0.16.0
httptools==0.6.4
humanfriendly==10.0
idna==3.10
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
Mako==1.3.10
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.6.2
numpy==2.3.1
packaging==25.0
pandas==2.3.0
pendulum==3.1.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==4.25.8
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.13.2
pymongoarrow==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
questionary==2.1.0
referencing==0.36.2
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.25.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.47.1
structlog==25.4.0
tabulate==0.9.0
termcolor==2.3.0
tomli==2.2.1
tomlkit==0.13.2
toposort==1.10
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
universal_pathlib==0.2.6
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchdog==5.0.3
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
yarl==1.20.1
yaspin==3.1.0
